pyproject.toml
[tool.poetry]
name = "test-b2b-task"
version = "0.1.0"
description = ""
authors = ["Anton <anton.Lezhneu@acg.group>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.1"
djangorestframework = "^3.15.2"
djangorestframework-jsonapi = "^7.0.2"
openapi = "^2.0.0"
django-filter = "^24.3"
mysql-connector-python = "^9.0.0"
drf-yasg = "^1.21.7"
python-decouple = "^3.8"
gunicorn = "^23.0.0"
sqlalchemy = "^2.0.32"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
ruff = "^0.6.2"
pytest-django = "^4.8.0"
mixer = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]